Setting:"폴더 경로 및 폴더 명 수정"
