+++ src/app/_components/_common/Loading/type.ts
