+++ src/app/_components/domain/main/works/sections/monthly/popular/Popular.tsx
@@ -1,0 +1,18 @@
+'use client';
+
+import { getPopularWorksApi } from '@/api/works';
+import WorkSection from '../../WorkSection';
+
+const Popular = () => {
+  return (
+    <WorkSection
+      title='이달의 인기 작품'
+      queryKey='popular'
+      type={'work'}
+      fetchFn={getPopularWorksApi}
+      variant='monthly'
+    />
+  );
+};
+
+export default Popular;
