+++ src/app/_components/domain/main/works/sections/recommend/webtoon/Webtoon.tsx
@@ -1,0 +1,19 @@
+'use client';
+import { getWebtoonWorksApi } from '@/api/works';
+import WorkSection from '../../WorkSection';
+
+const Webtoon = () => {
+  return (
+    <WorkSection
+      title='추천 웹툰'
+      queryKey='webtoon'
+      fetchFn={getWebtoonWorksApi}
+      type={'work'}
+      variant='webtoon'
+      hasMoreLink
+      keyword='webtoon'
+    />
+  );
+};
+
+export default Webtoon;
