--- src/app/components/domain/main/works/sections/best/Best.tsx
@@ -1,18 +1,0 @@
-'use client';
-
-import { getBestWorksApi } from '@/app/api/works';
-import WorkSection from '../WorkSection';
-
-const Best = () => {
-  return (
-    <WorkSection
-      title='감성코어 인기 best 작품 순위'
-      queryKey='todayBest'
-      type={'work'}
-      fetchFn={getBestWorksApi}
-      variant='best'
-    />
-  );
-};
-
-export default Best;
