--- src/app/components/domain/main/works/sections/monthly/author/Author.tsx
@@ -1,17 +1,0 @@
-'use client';
-import { getBestAuthorApi } from '@/app/api/works';
-import WorkSection from '../../WorkSection';
-
-const Author = () => {
-  return (
-    <WorkSection
-      title='이달의 우수 작가'
-      queryKey='author'
-      type={'author'}
-      fetchFn={getBestAuthorApi}
-      variant='monthly'
-    />
-  );
-};
-
-export default Author;
