--- src/app/components/domain/main/works/sections/recommend/poem/Poem.tsx
@@ -1,20 +1,0 @@
-'use client';
-
-import { getPomeWorksApi } from '@/app/api/works';
-import WorkSection from '../../WorkSection';
-
-const Poem = () => {
-  return (
-    <WorkSection
-      title='추천 시'
-      queryKey='poem'
-      fetchFn={getPomeWorksApi}
-      type={'work'}
-      variant='poem'
-      hasMoreLink
-      keyword='poem'
-    />
-  );
-};
-
-export default Poem;
