--- src/app/components/domain/main/works/sections/type.ts
@@ -1,14 +1,0 @@
-import { VariantProps } from 'class-variance-authority';
-import { sectionStyles } from './WorkSection';
-import { WorkData } from '@/app/api/works/type';
-import { CardVariant } from '@/app/components/_common/Cards/type';
-
-export interface WorkSectionProps extends VariantProps<typeof sectionStyles> {
-  title: string;
-  queryKey: string;
-  fetchFn: () => Promise<WorkData[]>;
-  type: 'work' | 'author';
-  hasMoreLink?: boolean;
-  variant: CardVariant;
-  keyword?: string;
-}
