+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import Best from './components/domain/main/works/sections/best/Best';
-import Recommend from './components/domain/main/works/sections/recommend/Recommend';
-import Monthly from './components/domain/main/works/sections/monthly/Monthly';
-import Tags from './components/domain/main/tags/Tags';
+import Best from './_components/domain/main/works/sections/best/Best';
+import Recommend from './_components/domain/main/works/sections/recommend/Recommend';
+import Monthly from './_components/domain/main/works/sections/monthly/Monthly';
+import Tags from './_components/domain/main/tags/Tags';
 
 const page = () => {
   return (
